[refactor] 느린 일기 페이지 리팩토링 (#218)
* style: style 파일 들여쓰기 수정

* style: 배경색 theme 활용하여 적용

* style: 공통된 부분을 공통 style 속성으로 관리

* style: theme font 속성으로 lettering-space 추가

* style: 주석 삭제 및 format date 변수명 변경

* refactor: 제목과 내용 blob 변환 삭제

* refactor: 파싱 함수 삭제

* refactor: api request 값을 위한  formdata 코드를 post와 patch 조건문 밖으로 빼내어 코드 간결화

* style: 세미콜론 및 들여쓰기 수정

* style: 들여쓰기 수정

* style: export문을 마지막에 작성

* refactor: useFileReader import문 수정

* refactor: DiaryForm 폴더 및 파일 생성후 Form 관리 컴포넌트 분리
