+++ src/hooks/common/useFileReader.jsx
@@ -1,12 +1,13 @@
-export function useFileReader(){
-    
-    const readData = (file, setFile) => {
-        const reader = new FileReader;
-        reader.readAsDataURL(file);
-        reader.onloadend = () => {
-            setFile(reader.result)
-        }
-    }
+function useFileReader(){
 
-    return readData;
+  const readData = (file, setFile) => {
+    const reader = new FileReader;
+    reader.readAsDataURL(file);
+    
+    reader.onloadend = () => {
+      setFile(reader.result)
+    }
+  };
+  return readData;
 }
+export default useFileReader;
