+++ src/styles/theme.js
@@ -48,6 +48,7 @@ const fonts = {
     font-style: normal;
     font-weight: 700;
     line-height: 2.2rem;
+    letter-spacing: -0.0408rem;
   `,
   title: css`
     font-family: 'Pretendard';
@@ -55,6 +56,7 @@ const fonts = {
     font-style: normal;
     font-weight: 700;
     line-height: 3rem;
+    letter-spacing: -0.0408rem;
   `,
   heading_01: css`
     font-family: 'Pretendard';
@@ -62,6 +64,7 @@ const fonts = {
     font-style: normal;
     font-weight: 600;
     line-height: 3rem;
+    letter-spacing: -0.0408rem;
   `,
   heading_03: css`
     font-family: 'Pretendard';
@@ -69,6 +72,7 @@ const fonts = {
     font-style: normal;
     font-weight: 700;
     line-height: 3rem;
+    letter-spacing: -0.0408rem;
   `,
 
   body_01: css`
@@ -77,6 +81,7 @@ const fonts = {
     font-style: normal;
     font-weight: 700;
     line-height: 2.2rem;
+    letter-spacing: -0.0408rem;
   `,
   body_02: css`
     font-family: 'Pretendard';
@@ -84,6 +89,7 @@ const fonts = {
     font-style: normal;
     font-weight: 400;
     line-height: 2.2rem;
+    letter-spacing: -0.0408rem;
   `,
   body_06: css`
     font-family: 'Pretendard';
@@ -91,6 +97,7 @@ const fonts = {
     font-style: normal;
     font-weight: 500;
     line-height: 2.2rem;
+    letter-spacing: -0.0408rem;
   `,
   body_08: css`
     font-family: 'Pretendard';
@@ -98,6 +105,7 @@ const fonts = {
     font-style: normal;
     font-weight: 400;
     line-height: 2.2rem;
+    letter-spacing: -0.0408rem;
   `,
   body_10: css`
     font-family: 'Pretendard';
@@ -105,6 +113,7 @@ const fonts = {
     font-style: normal;
     font-weight: 400;
     line-height: 2.2rem;
+    letter-spacing: -0.0408rem;
   `,
 
   caption_01: css`
@@ -113,6 +122,7 @@ const fonts = {
     font-style: normal;
     font-weight: 700;
     line-height: 1.6rem;
+    letter-spacing: -0.0408rem;
   `,
   caption_02: css`
     font-family: 'Pretendard';
@@ -120,6 +130,7 @@ const fonts = {
     font-style: normal;
     font-weight: 400;
     line-height: 1.6rem;
+    letter-spacing: -0.0408rem;
   `,
   caption_04: css`
     font-family: 'Pretendard';
@@ -127,6 +138,7 @@ const fonts = {
     font-style: normal;
     font-weight: 400;
     line-height: 2.2rem;
+    letter-spacing: -0.0408rem;
   `,
 };
